+++ luatest/init.lua
@@ -20,15 +20,10 @@ luatest.tarantool = require('luatest.tarantool')
 -- @see luatest.helpers
 luatest.helpers = require('luatest.helpers')
 
---- Class to manage tarantool instances.
+--- Class to manage Tarantool instances.
 --
 -- @see luatest.server
 luatest.Server = require('luatest.server')
-
---- Class to manage Tarantool instances, version 2.
---
--- @see luatest.server_v2
-luatest.ServerV2 = require('luatest.server_v2')
 
 local Group = require('luatest.group')
 local hooks = require('luatest.hooks')